Integrated scraped sites api with frontend

+++ backend/app/service/scraped_site_service.py
@@ -3,6 +3,7 @@ from app.model import db, ScrapedSite, ScrapedSiteSettings, JobListing
 from app.utils.scraper.scrape import getAllJobListings
 from app.utils.scraper.url_builder import ausgradUrlBuilder, seekUrlBuilder
 from app.utils.scraper.helper import findNewJobListings
+from app.utils.scraper.constants import GRAD_CONNECTION, SEEK
 
 from datetime import datetime, timezone
 
@@ -12,7 +13,7 @@ def get_all_scraped_sites():
     #first time db is empty -> need to create scraped sites
     if (len(scrapedSites) == 0):
         #create scraped sites
-        sites_to_scrape = ['Grad Connection', 'Seek']
+        sites_to_scrape = [GRAD_CONNECTION, SEEK]
         for site in sites_to_scrape:
             #create scraped site settings object with default settings
             scrapedSiteSettings = ScrapedSiteSettings(
@@ -21,14 +22,15 @@ def get_all_scraped_sites():
                 is_notification_enabled=True,
                 is_notify_email=True,
                 is_notify_notification=True,
+                max_pages_to_scrape=2,
                 search_keyword="software engineer",
             )
 
-            if (site == 'Grad Connection'):
+            if (site == GRAD_CONNECTION):
                 scrapedSiteSettings.location = "australia"
                 scrapedSiteSettings.job_type = ""
                 scrapedSiteSettings.classification = "engineering-software"
-            elif (site == 'Seek'):
+            elif (site == SEEK):
                 scrapedSiteSettings.location = "All Australia"
                 scrapedSiteSettings.work_type = ""
                 scrapedSiteSettings.classification = "information-communication-technology"
@@ -63,35 +65,51 @@ def scrape_site(scrape_site_id):
     if scrapedSiteSettings is None:
         return None
     
-    if (scrapedSite.website_name == "Grad Connection"):
+    if (scrapedSite.website_name == GRAD_CONNECTION):
         search_url = ausgradUrlBuilder(scrapedSiteSettings.search_keyword, scrapedSiteSettings.job_type, scrapedSiteSettings.classification, scrapedSiteSettings.location)
-    elif (scrapedSite.website_name == "Seek"):
+    elif (scrapedSite.website_name == SEEK):
         search_url = seekUrlBuilder(scrapedSiteSettings.search_keyword, scrapedSiteSettings.work_type, scrapedSiteSettings.classification, scrapedSiteSettings.location)
    
+    # scrape site
     scraped_jobs = getAllJobListings(scrapedSite.website_name, search_url, scrapedSiteSettings.max_pages_to_scrape)
+
+    # update scraped site last scrape date to db
+    scrapedSite.last_scrape_date = datetime.now(timezone.utc)
 
     # get existing job listings from db
     existing_job_listings = JobListing.query.filter_by(scraped_site_id=scrapedSite.id).all()
 
     # find new job listings
-    new_job_listings = findNewJobListings(scraped_jobs, existing_job_listings)
+    [updated_job_listings, found_new_jobs] = findNewJobListings(existing_job_listings, scraped_jobs)
+
+    if (not found_new_jobs):
+        #save changes of scraped site to db
+        db.session.commit()
+        return [existing_job_listing.to_dict() for existing_job_listing in existing_job_listings]
 
     # remove all old job listings
     for job_listing in existing_job_listings:
         db.session.delete(job_listing)
 
-    # add new job listings
-    for job_listing in new_job_listings:
+    result = []
+
+    # add new job listings to db
+    for job_listing in updated_job_listings:
+        job_listing = JobListing(
+            scraped_site_id=scrapedSite.id,
+            job_title=job_listing['job_title'],
+            company_name=job_listing['company_name'],
+            location=job_listing['location'],
+            job_description=job_listing['job_description'],
+            additional_info=job_listing['additional_info'],
+            salary=job_listing['salary'],
+            job_url=job_listing['job_url'],
+            job_date=job_listing['job_date'],
+            is_new=job_listing['is_new']
+        )
+        result.append(job_listing.to_dict())
         db.session.add(job_listing)
 
-    # update last scrape date
-    scrapedSite.last_scrape_date = datetime.now(timezone.utc)
-
     db.session.commit()
-
-    # website = "Grad Connection"
-    # # search_url = "https://seek.com.au/Software-Engineer-jobs/in-All-Sydney-NSW?sortmode=ListedDate"
-    # search_url = "https://au.gradconnection.com/internships/sydney/?title=Software+Engineer&ordering=-recent_job_created"
-    # scraped_jobs = getAllJobListings(website, search_url, 2)
-    return [job_listing.to_dict() for job_listing in new_job_listings]
+    return result
